Crate cairo

Bindings and wrappers for Cairo

Modules

enums

ffi

glib

prelude

Structs

Context

FontExtents

FontFace
FontFace represents a particular font at a particular weight, slant, and other characteristic but no size, transformation, or size.

FontOptions

Glyph

ImageSurface

LinearGradient

Matrix

Mesh

Path
Paths are the most basic drawing tools and are primarily used to implicitly generate simple masks.

PathSegments

RadialGradient

Rectangle

RectangleInt

RectangleVec

ScaledFont
ScaledFont represents a realization of a font face at a particular size and transformation and a certain set of font options.

SolidPattern

Surface

SurfacePattern

TextCluster

TextExtents

Enums

Antialias
Specifies the type of antialiasing to do when rendering text or shapes.

FillRule
FillRule is used to select how paths are filled. For both fill rules, whether or not a point is included in the fill is determined by taking a ray from that point to infinity and looking at intersections with the path. The ray can be in any direction, as long as it doesn't pass through the end point of a segment or have a tricky intersection such as intersecting tangent to the path. (Note that filling is not actually implemented in this way. This is just a description of the rule that is applied.)

Format

LineCap
Specifies how to render the endpoints of the path when stroking.

LineJoin
Specifies how to render the junction of two lines when stroking.

MeshCorner

Operator
Operator is used to set the compositing operator for all cairo drawing operations.

PathDataType
PathData is used to describe the type of one portion of a path when represented as a Path. See PathData for details.

PathSegment

Status
Status is used to indicate errors that can occur when using Cairo. In some cases it is returned directly by functions, but when using Context, the last error, if any, is stored in the context and can be retrieved with Context::status().

SurfaceType

Traits

Gradient

MatrixTrait

Pattern