Enum cairo::Operator

pub enum Operator {
    Clear,
    Source,
    Over,
    In,
    Out,
    Atop,
    Dest,
    DestOver,
    DestIn,
    DestOut,
    DestAtop,
    Xor,
    Add,
    Saturate,
    Multiply,
    Screen,
    Overlay,
    Darken,
    Lighten,
    ColorDodge,
    ColorBurn,
    HardLight,
    SoftLight,
    Difference,
    Exclusion,
    HslHue,
    HslSaturation,
    HslColor,
    HslLuminosity,
    // some variants omitted
}

Operator is used to set the compositing operator for all cairo drawing operations.

The default operator is Over.

The operators marked as unbounded modify their destination even outside of the mask layer (that is, their effect is not bound by the mask layer). However, their effect can still be limited by way of clipping.

To keep things simple, the operator descriptions here document the behavior for when both source and destination are either fully transparent or fully opaque. The actual implementation works for translucent layers too. For a more detailed explanation of the effects of each operator, including the mathematical definitions, see http:cairographics.org/operators/.

Variants
Clear

clear destination layer (bounded) (Since 1.0)

Source

replace destination layer (bounded) (Since 1.0)

Over

draw source layer on top of destination layer (bounded) (Since 1.0)

In

draw source where there was destination content (unbounded) (Since 1.0)

Out

draw source where there was no destination content (unbounded) (Since 1.0)

Atop

draw source on top of destination content and only there (Since 1.0)

Dest

ignore the source (Since 1.0)

DestOver

draw destination on top of source (Since 1.0)

DestIn

leave destination only where there was source content (unbounded) (Since 1.0)

DestOut

leave destination only where there was no source content (Since 1.0)

DestAtop

leave destination on top of source content and only there (unbounded) (Since 1.0)

Xor

source and destination are shown where there is only one of them (Since 1.0)

Add

source and destination layers are accumulated (Since 1.0)

Saturate

like over, but assuming source and dest are disjoint geometries (Since 1.0)

Multiply

source and destination layers are multiplied. This causes the result to be at least as dark as the darker inputs. (Since 1.10)

Screen

source and destination are complemented and multiplied. This causes the result to be at least as light as the lighter inputs. (Since 1.10)

Overlay

multiplies or screens, depending on the lightness of the destination color. (Since 1.10)

Darken

replaces the destination with the source if it is darker, otherwise keeps the source. (Since 1.10)

Lighten

replaces the destination with the source if it is lighter, otherwise keeps the source. (Since 1.10)

ColorDodge

brightens the destination color to reflect the source color. (Since 1.10)

ColorBurn

darkens the destination color to reflect the source color. (Since 1.10)

HardLight

Multiplies or screens, dependent on source color. (Since 1.10)

SoftLight

Darkens or lightens, dependent on source color. (Since 1.10)

Difference

Takes the difference of the source and destination color. (Since 1.10)

Exclusion

Produces an effect similar to difference, but with lower contrast. (Since 1.10)

HslHue

Creates a color with the hue of the source and the saturation and luminosity of the target. (Since 1.10)

HslSaturation

Creates a color with the saturation of the source and the hue and luminosity of the target. Painting with this mode onto a gray area produces no change. (Since 1.10)

HslColor

Creates a color with the hue and saturation of the source and the luminosity of the target. This preserves the gray levels of the target and is useful for coloring monochrome images or tinting color images. (Since 1.10)

HslLuminosity