Enum cairo::Status

pub enum Status {
    Success,
    NoMemory,
    InvalidRestore,
    InvalidPopGroup,
    NoCurrentPoint,
    InvalidMatrix,
    InvalidStatus,
    NullPointer,
    InvalidString,
    InvalidPathData,
    ReadError,
    WriteError,
    SurfaceFinished,
    SurfaceTypeMismatch,
    PatternTypeMismatch,
    InvalidContent,
    InvalidFormat,
    InvalidVisual,
    FileNotFound,
    InvalidDash,
    InvalidDscComment,
    InvalidIndex,
    ClipNotRepresentable,
    TempFileError,
    InvalidStride,
    FontTypeMismatch,
    UserFontImmutable,
    UserFontError,
    NegativeCount,
    InvalidClusters,
    InvalidSlant,
    InvalidWeight,
    InvalidSize,
    UserFontNotImplemented,
    DeviceTypeMismatch,
    DeviceError,
    InvalidMeshConstruction,
    DeviceFinished,
    LastStatus,
}

Status is used to indicate errors that can occur when using Cairo. In some cases it is returned directly by functions, but when using Context, the last error, if any, is stored in the context and can be retrieved with Context::status().

New entries may be added in future versions. Use Context::status_to_string() to get a human-readable representation of an error message.

Variants

Success

no error has occurred (Since 1.0)

NoMemory

out of memory (Since 1.0)

InvalidRestore

Context::restore() called without matching Context::save() (Since 1.0)

InvalidPopGroup

no saved group to pop, i.e. Context::pop_group() without matching Context::push_group() (Since 1.0)

NoCurrentPoint

no current point defined (Since 1.0)

InvalidMatrix

invalid matrix (not invertible) (Since 1.0)

InvalidStatus

invalid value for an input Status (Since 1.0)

NullPointer

NULL pointer (Since 1.0)

InvalidString

input string not valid UTF-8 (Since 1.0)

InvalidPathData

input path data not valid (Since 1.0)

ReadError

error while reading from input stream (Since 1.0)

WriteError

error while writing to output stream (Since 1.0)

SurfaceFinished

target surface has been finished (Since 1.0)

SurfaceTypeMismatch

the surface type is not appropriate for the operation (Since 1.0)

PatternTypeMismatch

the pattern type is not appropriate for the operation (Since 1.0)

InvalidContent

invalid value for an input Content (Since 1.0)

InvalidFormat

invalid value for an input Format (Since 1.0)

InvalidVisual

invalid value for an input Visual* (Since 1.0)

FileNotFound

file not found (Since 1.0)

InvalidDash

invalid value for a dash setting (Since 1.0)

InvalidDscComment

invalid value for a DSC comment (Since 1.2)

InvalidIndex

invalid index passed to getter (Since 1.4)

ClipNotRepresentable

clip region not representable in desired format (Since 1.4)

TempFileError

error creating or writing to a temporary file (Since 1.6)

InvalidStride

invalid value for stride (Since 1.6)

FontTypeMismatch

the font type is not appropriate for the operation (Since 1.8)

UserFontImmutable

the user-font is immutable (Since 1.8)

UserFontError

error occurred in a user-font callback function (Since 1.8)

NegativeCount

negative number used where it is not allowed (Since 1.8)

InvalidClusters

input clusters do not represent the accompanying text and glyph array (Since 1.8)

InvalidSlant

invalid value for an input FontSlant (Since 1.8)

InvalidWeight

invalid value for an input FontWeight (Since 1.8)

InvalidSize

invalid value (typically too big) for the size of the input (surface, pattern, etc.) (Since 1.10)

UserFontNotImplemented

user-font method not implemented (Since 1.10)

DeviceTypeMismatch

the device type is not appropriate for the operation (Since 1.10)

DeviceError

an operation to the device caused an unspecified error (Since 1.10)

InvalidMeshConstruction

a mesh pattern construction operation was used outside of a Context::mesh_pattern_begin_patch()/Context::mesh_pattern_end_patch() pair (Since 1.12)

DeviceFinished

target device has been finished (Since 1.12)

LastStatus

this is a special value indicating the number of status values defined in this enumeration. When using this value, note that the version of cairo at run-time may have additional status values defined than the value of this symbol at compile-time. (Since 1.10)

Methods

impl Status

fn ensure_valid(&self)