Enum cairo::enums::PatternType

pub enum PatternType {
    Solid,
    Surface,
    LinearGradient,
    RadialGradient,
    Mesh,
    RasterSource,
}

PatternType is used to describe the type of a given pattern.

The type of a pattern is determined by the function used to create it. The Pattern::create_rgb() and Pattern::create_rgba() functions create Solid patterns. The remaining cairo_pattern_create functions map to pattern types in obvious ways.

The pattern type can be queried with Pattern::get_type().

Most Pattern functions can be called with a pattern of any type, (though trying to change the extend or filter for a solid pattern will have no effect). A notable exception is Pattern::add_color_stop_rgb() and Pattern::add_color_stop_rgba() which must only be called with gradient patterns (either Linear or Radial). Otherwise the pattern will be shutdown and put into an error state.

New entries may be added in future versions.

Variants

Solid

The pattern is a solid (uniform) color. It may be opaque or translucent, since 1.2.

Surface

The pattern is a based on a surface (an image), since 1.2.

LinearGradient

The pattern is a linear gradient, since 1.2.

RadialGradient

The pattern is a radial gradient, since 1.2.

Mesh

The pattern is a mesh, since 1.12.

RasterSource

The pattern is a user pattern providing raster data, since 1.12.