Module cairo::glib

Modules

boxed

date
Date and Time Functions — calendrical calculations and miscellaneous time stuff

glib_container

object
Types that facilitate representing GObject descendants.

signal

source
Manages available sources of events for the main loop

traits

translate
Translation between GLib/GLib-based FFI types and their Rust counterparts.

types

Structs

Continue
Return type of idle and timeout functions.

Date

Elem

Error
The Error structure contains information about an error that has occurred.

List

ParamSpec

Permission

RevElem

SElem

SList

TimeVal
Represents a precise time, with seconds and microseconds. Similar to the struct timeval returned by the gettimeofday() UNIX system call.

TypeClass

TypeInstance

TypeInterface

Value
Possible improvement : store a function pointer inside the struct and make the struct templated

Enums

GDateMonth

GDateWeekday

ParamFlags

Type
A GLib or GLib-based library type

Traits

FFIGObject

GlibContainer

ValuePublic

Functions

idle_add
Adds a function to be called whenever there are no higher priority events pending to the default main loop.

timeout_add
Sets a function to be called at regular intervals, with the default priority, PRIORITY_DEFAULT.

timeout_add_seconds
Sets a function to be called at regular intervals with the default priority, PRIORITY_DEFAULT.

to_bool

to_gboolean

Type Definitions

AppInfo

GDateDay

GDateYear

GTime