Struct cairo::pdf::Writer

pub struct Writer<W: Write> { /* fields omitted */ }

A PDF surface that writes to a generic io::Write type (owning variant)

The Writer takes ownership of the write object. Once you're done using the surface, you can obtain the write object back using the finish() method.

If you would like the surface to reference the write object instead, use RefWriter.

Methods

impl<W: Write> Writer<W>

pub fn new(width: f64, height: f64, writer: W) -> Writer<W>

Create a new writer surface.

use std::fs::File;

use cairo::Context;
use cairo::pdf;
use cairo::prelude::*;

let surface = pdf::Writer::new(100.0, 100.0, File::create("test.pdf").unwrap());
let context = Context::new(&surface);

// Draw things on the context.

surface.finish();

pub fn writer(&self) -> &W

pub fn writer_mut(&mut self) -> &mut W

pub fn finish(self) -> W