Trait cairo::prelude::Pattern

pub trait Pattern {
    fn ensure_status(&self) { ... }
    fn status(&self) -> Status { ... }
    fn get_reference_count(&self) -> isize { ... }
    fn set_extend(&self, extend: Extend) { ... }
    fn get_extend(&self) -> Extend { ... }
    fn set_filter(&self, filter: Filter) { ... }
    fn get_filter(&self) -> Filter { ... }
}

Provided Methods

fn ensure_status(&self)

fn status(&self) -> Status

Checks whether an error has previously occurred for this pattern.

fn get_reference_count(&self) -> isize

Returns the current reference count of self.

fn set_extend(&self, extend: Extend)

Sets the mode to be used for drawing outside the area of a pattern. See cairo_extend_t for details on the semantics of each extend strategy.

The default extend mode is Extend::None for surface patterns and Extend::Pad for gradient patterns.

fn get_extend(&self) -> Extend

Gets the current extend mode for a pattern. See Extend enum for details on the semantics of each extend strategy.

fn set_filter(&self, filter: Filter)

Sets the filter to be used for resizing when using this pattern. See Filter enum for details on each filter.

Note that you might want to control filtering even when you do not have an explicit Pattern object, (for example when using cairo_set_source_surface()). In these cases, it is convenient to use cairo_get_source() to get access to the pattern that cairo creates implicitly. For example:

Context::set_source_surface(image, x, y);
p.set_filter(Filter::nearest);

fn get_filter(&self) -> Filter

Gets the current filter for a pattern. See Filter enum for details on each filter.