Module cairo::runner

This module represents a run of a Cairo program as a series of consecutive execution steps, each of which define the execution logic of Cairo instructions

Structs

CairoContext

A structure to store auxiliary variables throughout computation

CairoInstruction

This structure stores all the needed information relative to an instruction at a given step of computation

CairoProgram

This struct stores the needed information to run a program

CairoState

A structure to store program counter, allocation pointer and frame pointer

CairoStep

A data structure to store a current step of Cairo computation

Traits

Pointers

This trait contains functions to obtain the Cairo pointers (program counter, allocation pointer and frame pointer)