Struct cairo::Device

pub struct Device(_);

Implementations

impl Device

pub unsafe fn from_raw_none(ptr: *mut cairo_device_t) -> Device

pub unsafe fn from_raw_borrow(ptr: *mut cairo_device_t) -> Borrowed<Device>

pub unsafe fn from_raw_full(ptr: *mut cairo_device_t) -> Device

pub fn to_raw_none(&self) -> *mut cairo_device_t

pub fn create<P: AsRef<Path>>(filename: P) -> Option<Device>

pub fn from_recording_surface(
&self,
surface: &RecordingSurface
) -> Result<(), Error>

pub fn mode(&self) -> ScriptMode

pub fn set_mode(&self, mode: ScriptMode)

pub fn surface_create(
&self,
content: Content,
width: f64,
height: f64
) -> Result<Surface, Error>

pub fn surface_create_for_target(
&self,
target: &Surface
) -> Result<Surface, Error>

pub fn write_comment(&self, comment: &str)

pub fn finish(&self)

pub fn flush(&self)

pub fn type_(&self) -> DeviceType

pub fn acquire(&self) -> Result<DeviceAcquireGuard<'_>, Error>

pub fn observer_elapsed(&self) -> f64

pub fn observer_fill_elapsed(&self) -> f64

pub fn observer_glyphs_elapsed(&self) -> f64

pub fn observer_mask_elapsed(&self) -> f64

pub fn observer_paint_elapsed(&self) -> f64

pub fn observer_stroke_elapsed(&self) -> f64

pub fn debug_cap_xrender_version(&self, major_version: i32, minor_version: i32)

pub fn debug_get_precision(&self) -> i32

pub fn debug_set_precision(&self, precision: i32)

pub fn status(&self) -> Result<(), Error>

pub fn set_user_data<T: 'static>(
&self,
key: &'static UserDataKey<T>,
value: Rc<T>
) -> Result<(), Error>

Attach user data to self for the given key.

pub fn user_data<T: 'static>(
&self,
key: &'static UserDataKey<T>
) -> Option<Rc<T>>

Return the user data previously attached to self with the given key, if any.

pub fn user_data_ptr<T: 'static>(
&self,
key: &'static UserDataKey<T>
) -> Option<NonNull<T>>

Return the user data previously attached to self with the given key, if any, without incrementing the reference count.

The pointer is valid when it is returned from this method, until the cairo object that self represents is destroyed or remove_user_data or set_user_data is called with the same key.

pub fn remove_user_data<T: 'static>(
&self,
key: &'static UserDataKey<T>
) -> Result<(), Error>

Unattach from self the user data associated with key, if any. If there is no other Rc strong reference, the data is destroyed.

impl Device

pub fn connection(&self) -> XCBConnection

pub fn debug_cap_xshm_version(&self, major_version: i32, minor_version: i32)