Struct cairo::FontFace

pub struct FontFace(pub *mut cairo_font_face_t);

FontFace represents a particular font at a particular weight, slant, and other characteristic but no size, transformation, or size.

Font faces are created using font-backend-specific constructors, typically of the form Context::backend_font_face_create(), or implicitly using the toy text API by way of Context::select_font_face(). The resulting face can be accessed using Context::get_font_face().

Methods

impl FontFace

fn toy_create(family: &str, slant: FontSlant, weight: FontWeight) -> FontFace

Creates a font face from a triplet of family, slant, and weight. These font faces are used in implementation of the the cairo "toy" font API.

If family is the zero-length string "", the platform-specific default family is assumed. The default family then can be queried using FontFace::toy_get_family().

The Context::select_font_face() function uses this to create font faces. See that function for limitations and other details of toy font faces.

fn toy_get_family(&self) -> Option<String>

Gets the familly name of a toy font.

fn toy_get_slant(&self) -> FontSlant

Gets the slant a toy font.

fn toy_get_weight(&self) -> FontWeight

Gets the weight a toy font.

fn ensure_status(&self)

Checks whether an error has previously occurred for this font face.

fn get_type(&self) -> FontType

This function returns the type of the backend used to create a font face. See FontType for available types.

fn get_reference_count(&self) -> usize

Returns the current reference count of self.

fn reference(&self) -> FontFace

Increases the reference count on self by one. This prevents self from being destroyed until a matching call to FontFace drop trait is made.

The number of references to a FontFace can be get using FontFace::get_reference_count().