Struct cairo::FontOptions

pub struct FontOptions(_);

Methods

impl FontOptions

The font options specify how fonts should be rendered. Most of the time the font options implied by a surface are just right and do not need any changes, but for pixel-based targets tweaking font options may result in superior output on a particular display.

fn new() -> FontOptions

Allocates a new font options object with all options initialized to default values.

fn ensure_status(&self)

Checks whether an error has previously occurred for this font options object.

fn merge(&mut self, other: &mut FontOptions)

Merges non-default options from other into self, replacing existing values. This operation can be thought of as somewhat similar to compositing other onto self with the operation of Operator::Over.

fn hash(&self) -> u64

Compute a hash for the font options object; this value will be useful when storing an object containing a FontOptions in a hash table.

fn set_antialias(&self, antialias: Antialias)

Sets the antialiasing mode for the font options object. This specifies the type of antialiasing to do when rendering text.

fn get_antialias(&self) -> Antialias

Gets the antialiasing mode for the font options object.

fn set_subpixel_order(&self, order: SubpixelOrder)

Sets the subpixel order for the font options object. The subpixel order specifies the order of color elements within each pixel on the display device when rendering with an antialiasing mode of Antialias::Subpixel. See the documentation for SubpixelOrder for full details.

fn get_subpixel_order(&self) -> SubpixelOrder

Gets the subpixel order for the font options object. See the documentation for SubpixelOrder for full details.

fn set_hint_style(&self, hint_style: HintStyle)

Sets the hint style for font outlines for the font options object. This controls whether to fit font outlines to the pixel grid, and if so, whether to optimize for fidelity or contrast. See the documentation for HintStyle for full details.

fn get_hint_style(&self) -> HintStyle

Gets the hint style for font outlines for the font options object. See the documentation for HintStyle for full details.

fn set_hint_metrics(&self, hint_metrics: HintMetrics)

Sets the metrics hinting mode for the font options object. This controls whether metrics are quantized to integer values in device units. See the documentation for HintMetrics for full details.

fn get_hint_metrics(&self) -> HintMetrics

Gets the metrics hinting mode for the font options object. See the documentation for HintMetrics for full details.