Struct cairo::LinearGradient

pub struct LinearGradient { /* fields omitted */ }

Methods

impl LinearGradient

pub fn new(x0: f64, y0: f64, x1: f64, y1: f64) -> LinearGradient

Create a new linear gradient Pattern object along the line defined by (x0, y0) and (x1, y1). Before using the gradient pattern, a number of color stops should be defined using Pattern::add_color_stop_rgb() or Pattern::add_color_stop_rgba().

Note: The coordinates here are in pattern space. For a new pattern, pattern space is identical to user space, but the relationship between the spaces can be changed with Pattern::set_matrix().

pub fn get_linear_points(&self) -> (f64, f64, f64, f64)

Gets the gradient endpoints for a linear gradient.