Struct cairo::Matrix

#[repr(C)]pub struct Matrix {
    pub xx: c_double,
    pub yx: c_double,
    pub xy: c_double,
    pub yy: c_double,
    pub x0: c_double,
    pub y0: c_double,
}

A structure capable of holding a 4x4 matrix.

The contents of the Matrix structure are private and should never be accessed directly. A structure specifying a transformation between user-space coordinates and device coordinates. The transformation is given by

<programlisting> x_device = x_user * matrix->xx + y_user * matrix->xy + matrix->x0; y_device = x_user * matrix->yx + y_user * matrix->yy + matrix->y0; </programlisting>

Fields
xx: c_doubleyx: c_doublexy: c_doubleyy: c_doublex0: c_doubley0: c_double
Implementations

impl Matrix

pub fn identity() -> Matrix

pub fn new(xx: f64, yx: f64, xy: f64, yy: f64, x0: f64, y0: f64) -> Matrix

pub fn multiply(left: &Matrix, right: &Matrix) -> Matrix

Multiplies two Matrix.

Matrix multiplication is not commutative in general; the order of the factors matters. The product of this multiplication is (self × b)

b

a Matrix

res

return location for the matrix result

pub fn translate(&mut self, tx: f64, ty: f64)

Adds a translation transformation to self using the coordinates of the given Point3D.

This is the equivalent of calling Matrix::init_translate and then multiplying self with the translation matrix.

pos

a Point3D Changes the transformation represented by self to be the transformation given by first translating by (tx, ty) then applying the original transformation.

tx

amount to translate in the X direction

ty

amount to translate in the Y direction

pub fn scale(&mut self, sx: f64, sy: f64)

Adds a scaling transformation to self, using the three given factors.

This is the equivalent of calling Matrix::init_scale and then multiplying the matrix self with the scale matrix.

factor_x

scaling factor on the X axis

factor_y

scaling factor on the Y axis

factor_z

scaling factor on the Z axis Changes the transformation represented by self to be the transformation given by first scaling by sx in the X direction and sy in the Y direction then applying the original transformation.

scale_x

amount to scale by in X direction

scale_y

amount to scale by in Y direction

pub fn rotate(&mut self, angle: f64)

Adds a rotation transformation to self, using the given angle and axis vector.

This is the equivalent of calling Matrix::init_rotate and then multiplying the matrix self with the rotation matrix.

angle

the rotation angle, in degrees

axis

the rotation axis, as a Vec3 Changes the transformation represented by self to be the transformation given by first rotating by degrees degrees counter-clockwise then applying the original transformation.

degrees

degrees to rotate counter-clockwise

pub fn invert(&mut self)

pub fn try_invert(&self) -> Result<Matrix, Error>

pub fn transform_distance(&self, _dx: f64, _dy: f64) -> (f64, f64)

Transforms the distance vector (dx,dy) by self. This is similar to Matrix::transform_point except that the translation components of the transformation are ignored. The calculation of the returned vector is as follows:

<programlisting> dx2 = dx1 * xx + dy1 * xy; dy2 = dx1 * yx + dy1 * yy; </programlisting>

Affine transformations are position invariant, so the same vector always transforms to the same vector. If (x1,y1) transforms to (x2,y2) then (x1+dx1,y1+dy1) will transform to (x1+dx2,y1+dy2) for all values of x1 and x2.

dx

in/out X component of a distance vector

dy

in/out Y component of a distance vector

pub fn transform_point(&self, _x: f64, _y: f64) -> (f64, f64)

Transforms the given Point using the matrix self.

Unlike Matrix::transform_vec3, this function will take into account the fourth row vector of the Matrix when computing the dot product of each row vector of the matrix.

See also: graphene_simd4x4f_point3_mul

p

a Point

res

return location for the transformed Point Transforms the point (x, y) by self.

x

in/out X position

y

in/out Y position