Struct cairo::Mesh

pub struct Mesh {
    // some fields omitted
}

Methods

impl Mesh

fn wrap(pointer: *mut cairo_pattern_t) -> Mesh

fn reference(&self) -> Mesh

impl Mesh

fn new() -> Mesh

Create a new mesh pattern.

Mesh patterns are tensor-product patch meshes (type 7 shadings in PDF). Mesh patterns may also be used to create other types of shadings that are special cases of tensor-product patch meshes such as Coons patch meshes (type 6 shading in PDF) and Gouraud-shaded triangle meshes (type 4 and 5 shadings in PDF).

Mesh patterns consist of one or more tensor-product patches, which should be defined before using the mesh pattern. Using a mesh pattern with a partially defined patch as source or mask will put the context in an error status with a status of Status::InvalidMeshConstruction.

A tensor-product patch is defined by 4 Bézier curves (side 0, 1, 2, 3) and by 4 additional control points (P0, P1, P2, P3) that provide further control over the patch and complete the definition of the tensor-product patch. The corner C0 is the first point of the patch.

Degenerate sides are permitted so straight lines may be used. A zero length line on one side may be used to create 3 sided patches.

  C1     Side 1       C2
   +---------------+
   |               |
   |  P1       P2  |
   |               |

Side 0 | | Side 2 | | | | | P0 P3 | | | +---------------+ C0 Side 3 C3

Each patch is constructed by first calling Mesh::begin_patch(), then Mesh::move_to() to specify the first point in the patch (C0). Then the sides are specified with calls to Mesh::curve_to() and cairo_mesh_pattern_line_to().

The four additional control points (P0, P1, P2, P3) in a patch can be specified with Mesh::set_control_point().

At each corner of the patch (C0, C1, C2, C3) a color may be specified with Mesh::set_corner_color_rgb() or Mesh::set_corner_color_rgba(). Any corner whose color is not explicitly specified defaults to transparent black.

A Coons patch is a special case of the tensor-product patch where the control points are implicitly defined by the sides of the patch. The default value for any control point not specified is the implicit value for a Coons patch, i.e. if no control points are specified the patch is a Coons patch.

A triangle is a special case of the tensor-product patch where the control points are implicitly defined by the sides of the patch, all the sides are lines and one of them has length 0, i.e. if the patch is specified using just 3 lines, it is a triangle. If the corners connected by the 0-length side have the same color, the patch is a Gouraud-shaded triangle.

Patches may be oriented differently to the above diagram. For example the first point could be at the top left. The diagram only shows the relationship between the sides, corners and control points. Regardless of where the first point is located, when specifying colors, corner 0 will always be the first point, corner 1 the point between side 0 and side 1 etc.

Calling Mesh::end_patch() completes the current patch. If less than 4 sides have been defined, the first missing side is defined as a line from the current point to the first point of the patch (C0) and the other sides are degenerate lines from C0 to C0. The corners between the added sides will all be coincident with C0 of the patch and their color will be set to be the same as the color of C0.

Additional patches may be added with additional calls to Mesh::begin_patch()/Mesh::end_patch().

let mut pattern = Mesh::new();
 Add a Coons patch 
pattern.begin_patch();
pattern.move_to(0, 0);
pattern.curve_to(30, -30,  60,  30, 100, 0);
pattern.curve_to(60,  30, 130,  60, 100, 100);
pattern.curve_to(60,  70,  30, 130,   0, 100);
pattern.curve_to(30,  70, -30,  30,   0, 0);
pattern.set_corner_color_rgb(0, 1, 0, 0);
pattern.set_corner_color_rgb(1, 0, 1, 0);
pattern.set_corner_color_rgb(2, 0, 0, 1);
pattern.set_corner_color_rgb(3, 1, 1, 0);
pattern.end_patch();

 Add a Gouraud-shaded triangle 
pattern.begin_patch()
pattern.move_to(100, 100);
pattern.line_to(130, 130);
pattern.line_to(130,  70);
pattern.set_corner_color_rgb(0, 1, 0, 0);
pattern.set_corner_color_rgb(1, 0, 1, 0);
pattern.set_corner_color_rgb(2, 0, 0, 1);
pattern.end_patch();

When two patches overlap, the last one that has been added is drawn over the first one.

When a patch folds over itself, points are sorted depending on their parameter coordinates inside the patch. The v coordinate ranges from 0 to 1 when moving from side 3 to side 1; the u coordinate ranges from 0 to 1 when going from side 0 to side

Points with higher v coordinate hide points with lower v coordinate. When two points have the same v coordinate, the one with higher u coordinate is above. This means that points nearer to side 1 are above points nearer to side 3; when this is not sufficient to decide which point is above (for example when both points belong to side 1 or side 3) points nearer to side 2 are above points nearer to side 0.

For a complete definition of tensor-product patches, see the PDF specification (ISO32000), which describes the parametrization in detail.

Note: The coordinates are always in pattern space. For a new pattern, pattern space is identical to user space, but the relationship between the spaces can be changed with Pattern::set_matrix().

fn begin_patch(&self)

Begin a patch in a mesh pattern.

After calling this function, the patch shape should be defined with Mesh::move_to(), Mesh::line_to() and Mesh::curve_to().

After defining the patch, Mesh::end_patch() must be called before using pattern as a source or mask.

Note: If pattern is not a mesh pattern then pattern will be put into an error status with a status of Status::PatternTypeMismatch. If pattern already has a current patch, it will be put into an error status with a status of Status::InvalidMeshConstruction.

fn end_patch(&self)

Indicates the end of the current patch in a mesh pattern.

If the current patch has less than 4 sides, it is closed with a straight line from the current point to the first point of the patch as if Mesh::line_to() was used.

Note: If pattern is not a mesh pattern then pattern will be put into an error status with a status of Status::PatternTypeMismatch. If pattern has no current patch or the current patch has no current point, pattern will be put into an error status with a status of Status::InvalidMeshConstruction.

fn move_to(&self, x: f64, y: f64)

Define the first point of the current patch in a mesh pattern.

After this call the current point will be (x , y ).

Note: If pattern is not a mesh pattern then pattern will be put into an error status with a status of Status::PatternTypeMismatch. If pattern has no current patch or the current patch already has at least one side, pattern will be put into an error status with a status of Status::InvalidMeshConstruction.

fn line_to(&self, x: f64, y: f64)

Adds a line to the current patch from the current point to position (x , y ) in pattern-space coordinates.

If there is no current point before the call to cairo_mesh_pattern_line_to() this function will behave as Mesh::move_to(pattern , x , y ).

After this call the current point will be (x , y ).

Note: If pattern is not a mesh pattern then pattern will be put into an error status with a status of Status::PatternTypeMismatch. If pattern has no current patch or the current patch already has 4 sides, pattern will be put into an error status with a status of Status::InvalidMeshConstruction.

fn curve_to(&self, x1: f64, y1: f64, x2: f64, y2: f64, x3: f64, y3: f64)

Adds a cubic Bézier spline to the current patch from the current point to position (x3 , y3 ) in pattern-space coordinates, using (x1 , y1 ) and (x2 , y2 ) as the control points.

If the current patch has no current point before the call to Mesh::curve_to(), this function will behave as if preceded by a call to Mesh::move_to(pattern , x1 , y1 ).

After this call the current point will be (x3 , y3 ).

Note: If pattern is not a mesh pattern then pattern will be put into an error status with a status of Status::PatternTypeMismatch. If pattern has no current patch or the current patch already has 4 sides, pattern will be put into an error status with a status of Status::InvalidMeshConstruction.

fn set_control_point(&self, corner: MeshCorner, x: f64, y: f64)

Set an internal control point of the current patch.

Valid values for point_num are from 0 to 3 and identify the control points as explained in Mesh::new().

Note: If pattern is not a mesh pattern then pattern will be put into an error status with a status of Status::PatternTypeMismatch. If point_num is not valid, pattern will be put into an error status with a status of Status::InvalidIndex. If pattern has no current patch, pattern will be put into an error status with a status of Status::InvalidMeshConstruction.

fn get_control_point(&self, patch_num: usize, corner: MeshCorner) -> (f64, f64)

Gets the control point point_num of patch patch_num for a mesh pattern.

patch_num can range from 0 to n-1 where n is the number returned by Mesh::get_patch_count().

Valid values for point_num are from 0 to 3 and identify the control points as explained in Mesh::new().

fn set_corner_color_rgb(&self, corner: MeshCorner, red: f64, green: f64, blue: f64)

Sets the color of a corner of the current patch in a mesh pattern.

The color is specified in the same way as in Context::set_source_rgb().

Valid values for corner_num are from 0 to 3 and identify the corners as explained in Mesh::new().

Note: If pattern is not a mesh pattern then pattern will be put into an error status with a status of Status::PatternTypeMismatch. If corner_num is not valid, pattern will be put into an error status with a status of Status::InvalidIndex. If pattern has no current patch, pattern will be put into an error status with a status of Status::InvalidMeshConstruction.

fn set_corner_color_rgba(&self, corner: MeshCorner, red: f64, green: f64, blue: f64, alpha: f64)

Sets the color of a corner of the current patch in a mesh pattern.

The color is specified in the same way as in Context::set_source_rgba().

Valid values for corner_num are from 0 to 3 and identify the corners as explained in Mesh::new().

Note: If pattern is not a mesh pattern then pattern will be put into an error status with a status of Status::PatternTypeMismatch. If corner_num is not valid, pattern will be put into an error status with a status of Status::InvalidIndex. If pattern has no current patch, pattern will be put into an error status with a status of Status::InvalidMeshConstruction.

fn get_corner_color_rgba(&self, patch_num: usize, corner: MeshCorner) -> (f64, f64, f64, f64)

Gets the color information in corner corner_num of patch patch_num for a mesh pattern.

patch_num can range from 0 to n-1 where n is the number returned by Mesh::get_patch_count().

Valid values for corner_num are from 0 to 3 and identify the corners as explained in Mesh::new().

fn get_patch_count(&self) -> usize

Gets the number of patches specified in the given mesh pattern.

The number only includes patches which have been finished by calling Mesh::end_patch(). For example it will be 0 during the definition of the first patch.

fn get_path(&self, patch_num: usize) -> Path

Gets path defining the patch patch_num for a mesh pattern.

patch_num can range from 0 to n-1 where n is the number returned by Mesh::get_patch_count().