Struct cairo::SolidPattern

pub struct SolidPattern {
    // some fields omitted
}

Methods

impl SolidPattern

fn wrap(pointer: *mut cairo_pattern_t) -> SolidPattern

fn reference(&self) -> SolidPattern

impl SolidPattern

fn from_rgb(red: f64, green: f64, blue: f64) -> SolidPattern

Creates a new SolidPattern corresponding to an opaque color. The color components are floating point numbers in the range 0 to 1.

Note : If the values passed in are outside that range, they will be clamped.

fn from_rgba(red: f64, green: f64, blue: f64, alpha: f64) -> SolidPattern

Creates a new SolidPattern corresponding to a translucent color. The color components are floating point numbers in the range 0 to

Note : If the values passed in are outside that range, they will be clamped.

fn get_rgba(&self) -> (f64, f64, f64, f64)

Gets the solid color for a solid color pattern.