cairo
Struct Surface

pub struct Surface(/* private fields */);

Implementations

impl Surface

pub unsafe fn from_raw_none(ptr: *mut cairo_surface_t) -> Surface

pub unsafe fn from_raw_borrow(ptr: *mut cairo_surface_t) -> Borrowed<Surface>

pub unsafe fn from_raw_full(ptr: *mut cairo_surface_t) -> Result<Surface, Error>

pub fn to_raw_none(&self) -> *mut cairo_surface_t

pub fn create_similar( &self, content: Content, width: i32, height: i32, ) -> Result<Surface, Error>

pub fn create_for_rectangle(&self, bounds: Rectangle) -> Result<Surface, Error>

pub fn mime_data(&self, mime_type: &str) -> Option<Vec<u8>>

pub unsafe fn mime_data_raw(&self, mime_type: &str) -> Option<&[u8]>

pub fn set_mime_data<T: AsRef<[u8]> + 'static>( &self, mime_type: &str, slice: T, ) -> Result<(), Error>

pub fn supports_mime_type(&self, mime_type: &str) -> bool

pub fn device(&self) -> Option<Device>

pub fn content(&self) -> Content

pub fn set_device_offset(&self, x_offset: f64, y_offset: f64)

pub fn device_offset(&self) -> (f64, f64)

pub fn set_device_scale(&self, x_scale: f64, y_scale: f64)

pub fn device_scale(&self) -> (f64, f64)

pub fn set_fallback_resolution( &self, x_pixels_per_inch: f64, y_pixels_per_inch: f64, )

pub fn fallback_resolution(&self) -> (f64, f64)

pub fn create_similar_image( &self, format: Format, width: i32, height: i32, ) -> Result<ImageSurface, Error>

pub fn map_to_image( &self, extents: Option<RectangleInt>, ) -> Result<MappedImageSurface, Error>

pub fn mark_dirty(&self)

pub fn mark_dirty_rectangle(&self, x: i32, y: i32, width: i32, height: i32)

pub fn status(&self) -> Result<(), Error>

pub fn set_user_data<T: 'static>( &self, key: &'static UserDataKey<T>, value: Rc<T>, ) -> Result<(), Error>

Attach user data to self for the given key.

pub fn user_data<T: 'static>( &self, key: &'static UserDataKey<T>, ) -> Option<Rc<T>>

Return the user data previously attached to self with the given key, if any.

pub fn user_data_ptr<T: 'static>( &self, key: &'static UserDataKey<T>, ) -> Option<NonNull<T>>

Return the user data previously attached to self with the given key, if any, without incrementing the reference count.

The pointer is valid when it is returned from this method, until the cairo object that self represents is destroyed or remove_user_data or set_user_data is called with the same key.

pub fn remove_user_data<T: 'static>( &self, key: &'static UserDataKey<T>, ) -> Result<(), Error>

Unattached from self the user data associated with key, if any. If there is no other Rc strong reference, the data is destroyed.

impl Surface

pub fn flush(&self)

pub fn finish(&self)

pub fn type_(&self) -> SurfaceType

impl Surface

pub fn write_to_png<W: Write>(&self, stream: &mut W) -> Result<(), IoError>

This function writes the surface as a PNG image to the given stream.

If the underlying surface does not support being written as a PNG, this will return Error::SurfaceTypeMismatch

impl Surface

pub fn finish_output_stream(&self) -> Result<Box<dyn Any>, StreamWithError>

Finish the surface, then remove and return the output stream if any.

This calls Surface::finish, to make sure pending writes are done.

This is relevant for surfaces created for example with crate::PdfSurface::for_stream.

Use Box::downcast to recover the concrete stream type.

Panics

This method panics if:

This method was already called for this surface, or

This surface was not created with an output stream in the first place, or

A previous write to this surface panicked, or

A previous write happened while another write was ongoing, or

A write is ongoing now.

The latter two cases can only occur with a pathological output stream type that accesses the same surface again from Write::write_all.