Trait cairo::MatrixTrait

pub trait MatrixTrait {
    fn null() -> Matrix;

    fn new(xx: f64, yx: f64, xy: f64, yy: f64, x0: f64, y0: f64) -> Matrix;

    fn multiply(left: &Matrix, right: &Matrix) -> Matrix;

    fn identity() -> Matrix;

    fn init(&mut self, xx: f64, yx: f64, xy: f64, yy: f64, x0: f64, y0: f64);

    fn translate(&mut self, tx: f64, ty: f64);

    fn scale(&mut self, sx: f64, sy: f64);

    fn rotate(&mut self, angle: f64);

    fn invert(&mut self);

    fn try_invert(&self) -> Result<Matrix, Status>;

    fn transform_distance(&self, _dx: f64, _dy: f64) -> (f64, f64);

    fn transform_point(&self, _x: f64, _y: f64) -> (f64, f64);
}

Required Methods

fn null() -> Matrix

Creates a new Matrix filled with zeroes

fn new(xx: f64, yx: f64, xy: f64, yy: f64, x0: f64, y0: f64) -> Matrix

Creates a new matrix and fills it with given values

fn multiply(left: &Matrix, right: &Matrix) -> Matrix

Multiplies the affine transformations in a and b together and stores the result in the returned Matrix. The effect of the resulting transformation is to first apply the transformation in left to the coordinates and then apply the transformation in right to the coordinates.

It is allowable for the returned Matrix to be identical to either a or .

fn identity() -> Matrix

Returns a new matrix after modifying it to be an identity transformation.

fn init(&mut self, xx: f64, yx: f64, xy: f64, yy: f64, x0: f64, y0: f64)

Sets self to be the affine transformation given by xx , yx , xy , yy , x0 , y0. The transformation is given by:

x_new = xx * x + xy * y + x0;
y_new = yx * x + yy * y + y0;

fn translate(&mut self, tx: f64, ty: f64)

Applies a translation by tx, ty to the transformation in self. The effect of the new transformation is to first translate the coordinates by tx and ty, then apply the original transformation to the coordinates.

fn scale(&mut self, sx: f64, sy: f64)

Applies scaling by sx, sy to the transformation in self. The effect of the new transformation is to first scale the coordinates by sx and sy, then apply the original transformation to the coordinates.

fn rotate(&mut self, angle: f64)

Applies rotation by radians to the transformation in self. The effect of the new transformation is to first rotate the coordinates by radians , then apply the original transformation to the coordinates.

fn invert(&mut self)

Inverts a matrix in-place.

Panics

Panics if the matrix is not invertible: if the matrix collapses points together (it is degenerate), then it has no inverse and this function will panic.

Normally a non-invertible matrix indicates a bug; all matrices that originate from your code should always be valid, invertible matrices.

If you construct a matrix from untrusted data and need to validate it, you can use try_invert().

Example

use cairo::{Matrix, MatrixTrait};

let mut matrix = Matrix::identity();
matrix.invert();
assert!(matrix == Matrix::identity());

fn try_invert(&self) -> Result<Matrix, Status>

Tries to invert a matrix, and returns the inverted result or an error if the matrix is not invertible.

A matrix is not invertible if it collapses points together (it is degenerate).

Normally, matrices that originate from your code should always be valid, invertible matrices. You should use this function only to validate matrices that come from untrusted data.

Errors

Non-invertible matrices yield an Err(Status::InvalidMatrix) error.

Example

use cairo::{Matrix, MatrixTrait};

let matrix = Matrix::identity();
assert!(matrix.try_invert().unwrap() == Matrix::identity());

let all_zeros_matrix = Matrix::null();
assert!(all_zeros_matrix.try_invert().is_err());

fn transform_distance(&self, _dx: f64, _dy: f64) -> (f64, f64)

Transforms the distance vector (dx, dy) by self. This is similar to Matrix::transform_point() except that the translation components of the transformation are ignored. The calculation of the returned vector is as follows:

dx2 = dx1 * a + dy1 * c;
dy2 = dx1 * b + dy1 * d;

Affine transformations are position invariant, so the same vector always transforms to the same vector. If (x1 ,y1 ) transforms to (x2 ,y2 ) then (x1 +dx1 ,y1 +dy1 ) will transform to (x1 + dx2, y1 + dy2) for all values of x1 and x2 .

fn transform_point(&self, _x: f64, _y: f64) -> (f64, f64)

Transforms the point (x , y) by self.