Trait cairo::word::Decomposition

pub trait Decomposition<F> {

    fn off_dst(&self) -> F;

    fn off_op0(&self) -> F;

    fn off_op1(&self) -> F;

    fn flags(&self) -> Vec<F>;

    fn flag_at(&self, pos: usize) -> F;

    fn f_dst_fp(&self) -> F;

    fn f_op0_fp(&self) -> F;

    fn f_op1_val(&self) -> F;

    fn f_op1_fp(&self) -> F;

    fn f_op1_ap(&self) -> F;

    fn f_res_add(&self) -> F;

    fn f_res_mul(&self) -> F;

    fn f_pc_abs(&self) -> F;

    fn f_pc_rel(&self) -> F;

    fn f_pc_jnz(&self) -> F;

    fn f_ap_add(&self) -> F;

    fn f_ap_one(&self) -> F;

    fn f_opc_call(&self) -> F;

    fn f_opc_ret(&self) -> F;

    fn f_opc_aeq(&self) -> F;

    fn f15(&self) -> F;

    fn dst_reg(&self) -> u8;

    fn op0_reg(&self) -> u8;

    fn op1_src(&self) -> u8;

    fn res_log(&self) -> u8;

    fn pc_up(&self) -> u8;

    fn ap_up(&self) -> u8;

    fn opcode(&self) -> u8;

}

This trait contains methods that decompose a field element into CairoWord components

Required methods

fn off_dst(&self) -> F

Returns the destination offset in biased representation

fn off_op0(&self) -> F

Returns the first operand offset in biased representation

fn off_op1(&self) -> F

Returns the second operand offset in biased representation

fn flags(&self) -> Vec<F>

Returns vector of 16 flags

fn flag_at(&self, pos: usize) -> F

Returns i-th bit-flag

fn f_dst_fp(&self) -> F

Returns bit-flag for destination register as F

fn f_op0_fp(&self) -> F

Returns bit-flag for first operand register as F

fn f_op1_val(&self) -> F

Returns bit-flag for immediate value for second register as F

fn f_op1_fp(&self) -> F

Returns bit-flag for frame pointer for second register as F

fn f_op1_ap(&self) -> F

Returns bit-flag for allocation pointer for second regsiter as F

fn f_res_add(&self) -> F

Returns bit-flag for addition operation in right side as F

fn f_res_mul(&self) -> F

Returns bit-flag for multiplication operation in right side as F

fn f_pc_abs(&self) -> F

Returns bit-flag for program counter update being absolute jump as F

fn f_pc_rel(&self) -> F

Returns bit-flag for program counter update being relative jump as F

fn f_pc_jnz(&self) -> F

Returns bit-flag for program counter update being conditional jump as F

fn f_ap_add(&self) -> F

Returns bit-flag for allocation counter update being a manual addition as F

fn f_ap_one(&self) -> F

Returns bit-flag for allocation counter update being a self increment as F

fn f_opc_call(&self) -> F

Returns bit-flag for operation being a call as F

fn f_opc_ret(&self) -> F

Returns bit-flag for operation being a return as F

fn f_opc_aeq(&self) -> F

Returns bit-flag for operation being an assert-equal as F

fn f15(&self) -> F

Returns bit-flag for 16th position

fn dst_reg(&self) -> u8

Returns flagset for destination register

fn op0_reg(&self) -> u8

Returns flagset for first operand register

fn op1_src(&self) -> u8

Returns flagset for second operand register

fn res_log(&self) -> u8

Returns flagset for result logics

fn pc_up(&self) -> u8

Returns flagset for program counter update

fn ap_up(&self) -> u8

Returns flagset for allocation pointer update

fn opcode(&self) -> u8

Returns flagset for operation code