Trait cairo::word::FlagBits

pub trait FlagBits<F> {

    fn f_dst_fp(&self) -> F;
    fn f_op0_fp(&self) -> F;
    fn f_op1_val(&self) -> F;
    fn f_op1_fp(&self) -> F;
    fn f_op1_ap(&self) -> F;
    fn f_res_add(&self) -> F;
    fn f_res_mul(&self) -> F;
    fn f_pc_abs(&self) -> F;
    fn f_pc_rel(&self) -> F;
    fn f_pc_jnz(&self) -> F;
    fn f_ap_add(&self) -> F;
    fn f_ap_one(&self) -> F;
    fn f_opc_call(&self) -> F;
    fn f_opc_ret(&self) -> F;
    fn f_opc_aeq(&self) -> F;
    fn f15(&self) -> F;

}

This trait contains methods that decompose a field element into CairoWord flagbits

Required Methods

fn f_dst_fp(&self) -> F

Returns bit-flag for destination register as F

fn f_op0_fp(&self) -> F

Returns bit-flag for first operand register as F

fn f_op1_val(&self) -> F

Returns bit-flag for immediate value for second register as F

fn f_op1_fp(&self) -> F

Returns bit-flag for frame pointer for second register as F

fn f_op1_ap(&self) -> F

Returns bit-flag for allocation pointer for second regsiter as F

fn f_res_add(&self) -> F

Returns bit-flag for addition operation in right side as F

fn f_res_mul(&self) -> F

Returns bit-flag for multiplication operation in right side as F

fn f_pc_abs(&self) -> F

Returns bit-flag for program counter update being absolute jump as F

fn f_pc_rel(&self) -> F

Returns bit-flag for program counter update being relative jump as F

fn f_pc_jnz(&self) -> F

Returns bit-flag for program counter update being conditional jump as F

fn f_ap_add(&self) -> F

Returns bit-flag for allocation counter update being a manual addition as F

fn f_ap_one(&self) -> F

Returns bit-flag for allocation counter update being a self increment as F

fn f_opc_call(&self) -> F

Returns bit-flag for operation being a call as F

fn f_opc_ret(&self) -> F

Returns bit-flag for operation being a return as F

fn f_opc_aeq(&self) -> F

Returns bit-flag for operation being an assert-equal as F

fn f15(&self) -> F

Returns bit-flag for 16th position