Crate cairo

Cairo bindings

This library contains safe Rust bindings for Cairo. It is a part of Gtk-rs.

Crate features

Default-on features

use_glib - Use with glib

Fileformat features

png - Reading and writing PNG images

pdf - Rendering PDF documents

svg - Rendering SVG documents

ps - Rendering PostScript documents

Cairo API version features

v1_14 - Use Cairo 1.14 APIs

v1_16 - Use Cairo 1.16 APIs

Documentation features

embed-lgpl-docs - Embed API docs locally

purge-lgpl-docs - Remove API docs again (counterpart to embed-lgpl-docs)

dox - Used to keep system dependent items in documentation

X Window features

xcb - X Window System rendering using the XCB library

xlib - X Window System rendering using XLib

Windows API features

win32-surface - Microsoft Windows surface support

Structs

Context

Device

FontExtents

FontFace

FontOptions

Glyph

Gradient

ImageSurface

ImageSurfaceData

LinearGradient

MappedImageSurface

Matrix

Mesh

Path
Paths are the most basic drawing tools and are primarily used to implicitly generate simple masks.

PathSegments

Pattern

PdfOutline

RadialGradient

RecordingSurface

Rectangle

RectangleInt

RectangleList

Region

ScaledFont

SolidPattern

Surface

SurfacePattern

TextCluster

TextExtents

UserDataKey
A key for indexing user data in various cairo types.

Version

Enums

Antialias

BorrowError

Content

DeviceType

Extend

FillRule

Filter

FontSlant

FontType

FontWeight

Format

FtSynthesize

HintMetrics

HintStyle

IoError

LineCap

LineJoin

MeshCorner

Operator

PathDataType

PathSegment

PatternType

RegionOverlap

ScriptMode

Status

SubpixelOrder

SurfaceType

TextClusterFlags

Constants

MIME_TYPE_JP2

MIME_TYPE_JPEG

MIME_TYPE_PNG

MIME_TYPE_UNIQUE_ID

MIME_TYPE_URI

Functions

debug_reset_static_data⚠

get_version_string