Crate cairo

Cairo bindings

This library contains safe Rust bindings for Cairo. It is a part of Gtk-rs.

Crate features

Default-on features

use_glib - Use with glib

Fileformat features

png - Reading and writing PNG images

pdf - Rendering PDF documents

svg - Rendering SVG documents

ps - Rendering PostScript documents

Cairo API version features

v1_14 - Use Cairo 1.14 APIs

v1_16 - Use Cairo 1.16 APIs

Documentation features

embed-lgpl-docs - Embed API docs locally

purge-lgpl-docs - Remove API docs again (counterpart to embed-lgpl-docs)

dox - Used to keep system dependent items in documentation

X Window features

xcb - X Window System rendering using the XCB library

xlib - X Window System rendering using XLib

Windows API features

win32-surface - Microsoft Windows surface support

Re-exports

pub use ffi;

pub use freetype_crate as freetype;

pub use glib;

Structs

Context
The Context structure stores global information used to control the itemization process.

Device
The Device object represents a single input device, such as a keyboard, a mouse, a touchpad, etc.

FontExtents

FontFace

FontOptions

Glyph

Gradient

ImageSurface

ImageSurfaceData

LinearGradient

MappedImageSurface

Matrix
A structure capable of holding a 4x4 matrix.

Mesh

Path

PathSegments

Pattern

PdfOutline

PdfSurface

PsSurface

QuartzSurface

RadialGradient

RecordingSurface

Rectangle
A data structure for holding a rectangle. Those coordinates are relative to the component top-level parent. Defines the position and size of a rectangle. It is identical to cairo::RectangleInt.

RectangleInt

RectangleList

Region

ScaledFont

SolidPattern

StreamWithError

Surface

SurfacePattern

SvgSurface

TextCluster

TextExtents

UserDataKey

Version

Win32Surface

XCBConnection

XCBDrawable

XCBPixmap

XCBRenderPictFormInfo

XCBScreen

XCBSurface

XCBVisualType

Enums

Antialias

BorrowError

Content

DeviceType

Error

Extend

FillRule

Filter

FontSlant

FontType

FontWeight

Format

FtSynthesize

HintMetrics

HintStyle

IoError

LineCap

LineJoin

MeshCorner

Operator

PathDataType

PathSegment

PatternType

PdfMetadata

PdfVersion

PsLevel

RegionOverlap

ScriptMode

SubpixelOrder

SurfaceType

SvgUnit

SvgVersion

TextClusterFlags

Constants

CAIRO_TAG_DEST

CAIRO_TAG_LINK

MIME_TYPE_CCITT_FAX

MIME_TYPE_CCITT_FAX_PARAMS

MIME_TYPE_EPS

MIME_TYPE_EPS_PARAMS

MIME_TYPE_JBIG2

MIME_TYPE_JBIG2_GLOBAL

MIME_TYPE_JBIG2_GLOBAL_ID

MIME_TYPE_JP2

MIME_TYPE_JPEG

MIME_TYPE_PNG

MIME_TYPE_UNIQUE_ID

MIME_TYPE_URI

PDF_OUTLINE_ROOT

Functions

debug_reset_static_data⚠

get_version_string