Trait cairo::prelude::Gradient

pub trait Gradient: Pattern {
    fn add_color_stop_rgb(&self, offset: f64, red: f64, green: f64, blue: f64) { ... }
    fn add_color_stop_rgba(&self, offset: f64, red: f64, green: f64, blue: f64, alpha: f64) { ... }
    fn get_color_stop_count(&self) -> isize { ... }
    fn get_color_stop_rgba(&self, index: isize) -> (f64, f64, f64, f64, f64) { ... }
}

Provided Methods

fn add_color_stop_rgb(&self, offset: f64, red: f64, green: f64, blue: f64)

Adds an opaque color stop to a gradient pattern. The offset specifies the location along the gradient's control vector. For example, a linear gradient's control vector is from (x0,y0) to (x1,y1) while a radial gradient's control vector is from any point on the start circle to the corresponding point on the end circle.

The color is specified in the same way as in Context::set_source_rgba().

If two (or more) stops are specified with identical offset values, they will be sorted according to the order in which the stops are added, (stops added earlier will compare less than stops added later). This can be useful for reliably making sharp color transitions instead of the typical blend.

Note: If the pattern is not a gradient pattern, (eg. a linear or radial pattern), then the pattern will be put into an error status with a status of StatusPattern::TypeMismatch.

fn add_color_stop_rgba(&self, offset: f64, red: f64, green: f64, blue: f64, alpha: f64)

Adds a translucent color stop to a gradient pattern. The offset specifies the location along the gradient's control vector. For example, a linear gradient's control vector is from (x0,y0) to (x1,y1) while a radial gradient's control vector is from any point on the start circle to the corresponding point on the end circle.

The color is specified in the same way as in Context::set_source_rgba().

If two (or more) stops are specified with identical offset values, they will be sorted according to the order in which the stops are added, (stops added earlier will compare less than stops added later). This can be useful for reliably making sharp color transitions instead of the typical blend.

Note: If the pattern is not a gradient pattern, (eg. a linear or radial pattern), then the pattern will be put into an error status with a status of StatusPattern::TypeMismatch.

fn get_color_stop_count(&self) -> isize

Gets the number of color stops specified in the given gradient pattern.

fn get_color_stop_rgba(&self, index: isize) -> (f64, f64, f64, f64, f64)

Gets the color and offset information at the given index for a gradient pattern. Values of index range from 0 to n-1 where n is the number returned by Pattern::get_color_stop_count().