Struct cairo::Context

pub struct Context(_);

The Context structure stores global information used to control the itemization process.

Implementations

impl Context

pub unsafe fn from_raw_none(ptr: *mut cairo_t) -> Context

pub unsafe fn from_raw_borrow(ptr: *mut cairo_t) -> Borrowed<Context>

pub unsafe fn from_raw_full(ptr: *mut cairo_t) -> Context

pub fn to_raw_none(&self) -> *mut cairo_t

pub fn status(&self) -> Result<(), Error>

pub fn new(target: &Surface) -> Context

Creates a new Context initialized to default values.

This function is not particularly useful as it should always be followed by a Context::set_font_map call, and the function FontMapExt::create_context does these two steps together and hence users are recommended to use that.

If you are using Pango as part of a higher-level system, that system may have it's own way of create a Context. For instance, the GTK+ toolkit has, among others, gdk_pango_context_get_for_screen, and gtk_widget_get_pango_context. Use those instead.

Returns

the newly allocated Context, which should be freed with gobject::Object::unref.

pub fn save(&self) -> Result<(), Error>

pub fn restore(&self) -> Result<(), Error>

pub fn get_target(&self) -> Surface

pub fn push_group(&self)

pub fn push_group_with_content(&self, content: Content)

pub fn pop_group(&self) -> Pattern

pub fn pop_group_to_source(&self)

pub fn get_group_target(&self) -> Surface

pub fn set_source_rgb(&self, red: f64, green: f64, blue: f64)

pub fn set_source_rgba(&self, red: f64, green: f64, blue: f64, alpha: f64)

pub fn set_source(&self, source: &Pattern)

pub fn get_source(&self) -> Pattern

pub fn set_source_surface(&self, surface: &Surface, x: f64, y: f64)

pub fn set_antialias(&self, antialias: Antialias)

pub fn get_antialias(&self) -> Antialias

pub fn set_dash(&self, dashes: &[f64], offset: f64)

pub fn get_dash_count(&self) -> i32

pub fn get_dash(&self) -> (Vec<f64>, f64)

pub fn get_dash_dashes(&self) -> Vec<f64>

pub fn get_dash_offset(&self) -> f64

pub fn set_fill_rule(&self, fill_rule: FillRule)

pub fn get_fill_rule(&self) -> FillRule

pub fn set_line_cap(&self, arg: LineCap)

pub fn get_line_cap(&self) -> LineCap

pub fn set_line_join(&self, arg: LineJoin)

pub fn get_line_join(&self) -> LineJoin

pub fn set_line_width(&self, arg: f64)

pub fn get_line_width(&self) -> f64

pub fn set_miter_limit(&self, arg: f64)

pub fn get_miter_limit(&self) -> f64

pub fn set_operator(&self, op: Operator)

pub fn get_operator(&self) -> Operator

pub fn set_tolerance(&self, arg: f64)

pub fn get_tolerance(&self) -> f64

pub fn clip(&self)

pub fn clip_preserve(&self)

pub fn clip_extents(&self) -> (f64, f64, f64, f64)

pub fn in_clip(&self, x: f64, y: f64) -> bool

pub fn reset_clip(&self)

pub fn copy_clip_rectangle_list(&self) -> Result<RectangleList, Error>

pub fn fill(&self)

pub fn fill_preserve(&self)

pub fn fill_extents(&self) -> (f64, f64, f64, f64)

pub fn in_fill(&self, x: f64, y: f64) -> bool

pub fn mask(&self, pattern: &Pattern)

pub fn mask_surface(&self, surface: &Surface, x: f64, y: f64)

pub fn paint(&self)

pub fn paint_with_alpha(&self, alpha: f64)

pub fn stroke(&self)

pub fn stroke_preserve(&self)

pub fn stroke_extents(&self) -> (f64, f64, f64, f64)

pub fn in_stroke(&self, x: f64, y: f64) -> bool

pub fn copy_page(&self)

pub fn show_page(&self)

pub fn get_reference_count(&self) -> u32

pub fn translate(&self, tx: f64, ty: f64)

pub fn scale(&self, sx: f64, sy: f64)

pub fn rotate(&self, angle: f64)

pub fn transform(&self, matrix: Matrix)

pub fn set_matrix(&self, matrix: Matrix)

Sets the transformation matrix that will be applied when rendering with this context. Note that reported metrics are in the user space coordinates before the application of the matrix, not device-space coordinates after the application of the matrix. So, they don't scale with the matrix, though they may change slightly for different matrices, depending on how the text is fit to the pixel grid.

matrix

a Matrix, or None to unset any existing matrix. (No matrix set is the same as setting the identity matrix.)

pub fn get_matrix(&self) -> Matrix

Gets the transformation matrix that will be applied when rendering with this context. See Context::set_matrix.

Returns

the matrix, or None if no matrix has been set (which is the same as the identity matrix). The returned matrix is owned by Pango and must not be modified or freed.

pub fn identity_matrix(&self)

pub fn user_to_device(&self, mut x: f64, mut y: f64) -> (f64, f64)

pub fn user_to_device_distance(&self, mut dx: f64, mut dy: f64) -> (f64, f64)

pub fn device_to_user(&self, mut x: f64, mut y: f64) -> (f64, f64)

pub fn device_to_user_distance(&self, mut dx: f64, mut dy: f64) -> (f64, f64)

pub fn select_font_face(
&self,
family: &str,
slant: FontSlant,
weight: FontWeight
)

pub fn set_font_size(&self, size: f64)

pub fn set_font_matrix(&self, matrix: Matrix)

pub fn get_font_matrix(&self) -> Matrix

pub fn set_font_options(&self, options: &FontOptions)

pub fn get_font_options(&self) -> FontOptions

pub fn set_font_face(&self, font_face: &FontFace)

pub fn get_font_face(&self) -> FontFace

pub fn set_scaled_font(&self, scaled_font: &ScaledFont)

pub fn get_scaled_font(&self) -> ScaledFont

pub fn show_text(&self, text: &str)

pub fn show_glyphs(&self, glyphs: &[Glyph])

pub fn show_text_glyphs(
&self,
text: &str,
glyphs: &[Glyph],
clusters: &[TextCluster],
cluster_flags: TextClusterFlags
)

pub fn font_extents(&self) -> FontExtents

pub fn text_extents(&self, text: &str) -> TextExtents

pub fn glyph_extents(&self, glyphs: &[Glyph]) -> TextExtents

pub fn copy_path(&self) -> Path

pub fn copy_path_flat(&self) -> Path

pub fn append_path(&self, path: &Path)

pub fn has_current_point(&self) -> bool

pub fn get_current_point(&self) -> (f64, f64)

pub fn new_path(&self)

pub fn new_sub_path(&self)

pub fn close_path(&self)

pub fn arc(&self, xc: f64, yc: f64, radius: f64, angle1: f64, angle2: f64)

pub fn arc_negative(
&self,
xc: f64,
yc: f64,
radius: f64,
angle1: f64,
angle2: f64
)

pub fn curve_to(&self, x1: f64, y1: f64, x2: f64, y2: f64, x3: f64, y3: f64)

pub fn line_to(&self, x: f64, y: f64)

pub fn move_to(&self, x: f64, y: f64)

pub fn rectangle(&self, x: f64, y: f64, width: f64, height: f64)

pub fn text_path(&self, str_: &str)

pub fn glyph_path(&self, glyphs: &[Glyph])

pub fn rel_curve_to(
&self,
dx1: f64,
dy1: f64,
dx2: f64,
dy2: f64,
dx3: f64,
dy3: f64
)

pub fn rel_line_to(&self, dx: f64, dy: f64)

pub fn rel_move_to(&self, dx: f64, dy: f64)

pub fn path_extents(&self) -> (f64, f64, f64, f64)

pub fn tag_begin(&self, tag_name: &str, attributes: &str)

pub fn tag_end(&self, tag_name: &str)