Struct cairo::Device

pub struct Device(_, _);

Methods

impl Device

pub unsafe fn from_raw_none(ptr: *mut cairo_device_t) -> Device

pub unsafe fn from_raw_borrow(ptr: *mut cairo_device_t) -> Device

pub unsafe fn from_raw_full(ptr: *mut cairo_device_t) -> Device

pub fn to_raw_none(&self) -> *mut cairo_device_t

pub fn create<P: AsRef<Path>>(filename: P) -> Option<Device>

pub fn from_recording_surface(&self, surface: &RecordingSurface) -> Status

pub fn get_mode(&self) -> ScriptMode

pub fn set_mode(&self, mode: ScriptMode)

pub fn surface_create(
&self,
content: Content,
width: f64,
height: f64
) -> Result<Surface, Status>

pub fn surface_create_for_target(
&self,
target: &Surface
) -> Result<Surface, Status>

pub fn write_comment(&self, comment: &str)

pub fn status(&self) -> Status

pub fn finish(&self)

pub fn flush(&self)

pub fn get_type(&self) -> DeviceType

pub fn acquire(&self) -> Status

pub fn release(&self)

pub fn observer_elapsed(&self) -> f64

pub fn observer_fill_elapsed(&self) -> f64

pub fn observer_glyphs_elapsed(&self) -> f64

pub fn observer_mask_elapsed(&self) -> f64

pub fn observer_paint_elapsed(&self) -> f64

pub fn observer_stroke_elapsed(&self) -> f64

pub fn set_user_data<T: 'static>(
&self,
key: &'static UserDataKey<T>,
value: Rc<T>
)

Attach user data to self for the given key.

pub fn get_user_data<T: 'static>(
&self,
key: &'static UserDataKey<T>
) -> Option<Rc<T>>

Return the user data previously attached to self with the given key, if any.

pub fn get_user_data_ptr<T: 'static>(
&self,
key: &'static UserDataKey<T>
) -> Option<NonNull<T>>

Return the user data previously attached to self with the given key, if any, without incrementing the reference count.

The pointer is valid when it is returned from this method, until the cairo object that self represents is destroyed or remove_user_data or set_user_data is called with the same key.

pub fn remove_user_data<T: 'static>(&self, key: &'static UserDataKey<T>)

Unattach from self the user data associated with key, if any. If there is no other Rc strong reference, the data is destroyed.