Struct cairo::FontFace

pub struct FontFace(_);

Implementations

impl FontFace

pub fn toy_create(
family: &str,
slant: FontSlant,
weight: FontWeight
) -> Result<FontFace, Error>

pub unsafe fn create_from_ft(face: FT_Face) -> Result<FontFace, Error>

pub unsafe fn create_from_ft_with_flags(
face: FT_Face,
load_flags: c_int
) -> Result<FontFace, Error>

pub unsafe fn from_raw_full(ptr: *mut cairo_font_face_t) -> FontFace

pub unsafe fn from_raw_none(ptr: *mut cairo_font_face_t) -> FontFace

pub fn to_raw_none(&self) -> *mut cairo_font_face_t

pub fn toy_get_family(&self) -> Option<String>

pub fn toy_get_slant(&self) -> FontSlant

pub fn toy_get_weight(&self) -> FontWeight

pub fn type_(&self) -> FontType

pub fn reference_count(&self) -> usize

pub fn synthesize(&self) -> FtSynthesize

pub fn set_synthesize(&self, synth_flags: FtSynthesize)

pub fn unset_synthesize(&self, synth_flags: FtSynthesize)

pub fn status(&self) -> Result<(), Error>

pub fn set_user_data<T: 'static>(
&self,
key: &'static UserDataKey<T>,
value: Rc<T>
) -> Result<(), Error>

Attach user data to self for the given key.

pub fn user_data<T: 'static>(
&self,
key: &'static UserDataKey<T>
) -> Option<Rc<T>>

Return the user data previously attached to self with the given key, if any.

pub fn user_data_ptr<T: 'static>(
&self,
key: &'static UserDataKey<T>
) -> Option<NonNull<T>>

Return the user data previously attached to self with the given key, if any, without incrementing the reference count.

The pointer is valid when it is returned from this method, until the cairo object that self represents is destroyed or remove_user_data or set_user_data is called with the same key.

pub fn remove_user_data<T: 'static>(
&self,
key: &'static UserDataKey<T>
) -> Result<(), Error>

Unattach from self the user data associated with key, if any. If there is no other Rc strong reference, the data is destroyed.