Struct cairo::RadialGradient

pub struct RadialGradient { /* fields omitted */ }

Methods

impl RadialGradient

pub fn wrap(pointer: *mut cairo_pattern_t) -> RadialGradient

pub fn reference(&self) -> RadialGradient

impl RadialGradient

pub fn new(
x0: f64,
y0: f64,
r0: f64,
x1: f64,
y1: f64,
r1: f64
) -> RadialGradient

Creates a new radial gradient Pattern between the two circles defined by (cx0, cy0, radius0) and (cx1, cy1, radius1). Before using the gradient pattern, a number of color stops should be defined using Pattern::add_color_stop_rgb() or Pattern::add_color_stop_rgba().

Note: The coordinates here are in pattern space. For a new pattern, pattern space is identical to user space, but the relationship between the spaces can be changed with Pattern::set_matrix().

pub fn get_radial_circles(&self) -> (f64, f64, f64, f64)

Gets the gradient endpoint circles for a radial gradient, each specified as a center coordinate and a radius.