Struct cairo::ScaledFont

pub struct ScaledFont(pub *mut cairo_scaled_font_t);

ScaledFont represents a realization of a font face at a particular size and transformation and a certain set of font options.

Methods

impl ScaledFont

fn new(font_face: FontFace, font_matrix: &mut Matrix, ctm: &mut Matrix, options: FontOptions) -> ScaledFont

Creates a ScaledFont object from a font face and matrices that describe the size of the font and the environment in which it will be used.

fn ensure_status(&self)

Checks whether an error has previously occurred for this ScaledFont.

fn get_type(&self) -> FontType

This function returns the type of the backend used to create a scaled font. See FontType for available types. However, this function never returns FontType::Toy.

fn get_reference_count(&self) -> usize

Returns the current reference count of self.

fn reference(&self) -> ScaledFont

Increases the reference count on self by one. This prevents self from being destroyed until a matching call to ScaledFont drop trait is made.

The number of references to a cairo_scaled_font_t can be get using ScaledFont::get_reference_count().