Struct cairo::ScaledFont

#[repr(transparent)]
pub struct ScaledFont { /* private fields */ }

Implementations

impl ScaledFont

pub fn as_ptr(&self) -> *mut cairo_scaled_font_t

Return the inner pointer to the underlying C value.

impl ScaledFont

pub fn new(
font_face: &FontFace,
font_matrix: &Matrix,
ctm: &Matrix,
options: &FontOptions
) -> Result<ScaledFont, Error>

pub fn to_raw_none(&self) -> *mut cairo_scaled_font_t

pub unsafe fn from_raw_full(ptr: *mut cairo_scaled_font_t) -> ScaledFont

pub unsafe fn from_raw_none(ptr: *mut cairo_scaled_font_t) -> ScaledFont

pub fn type_(&self) -> FontType

pub fn reference_count(&self) -> usize

pub fn extents(&self) -> FontExtents

pub fn text_extents(&self, text: &str) -> TextExtents

pub fn glyph_extents(&self, glyphs: &[Glyph]) -> TextExtents

pub fn text_to_glyphs(
&self,
x: f64,
y: f64,
text: &str
) -> Result<(Vec<Glyph>, Vec<TextCluster>), Error>

pub fn font_face(&self) -> FontFace

pub fn font_options(&self) -> Result<FontOptions, Error>

pub fn font_matrix(&self) -> Matrix

pub fn ctm(&self) -> Matrix

pub fn scale_matrix(&self) -> Matrix

pub fn status(&self) -> Result<(), Error>

pub fn set_user_data<T: 'static>(
&self,
key: &'static UserDataKey<T>,
value: Rc<T>
) -> Result<(), Error>

Attach user data to self for the given key.

pub fn user_data<T: 'static>(
&self,
key: &'static UserDataKey<T>
) -> Option<Rc<T>>

Return the user data previously attached to self with the given key, if any.

pub fn user_data_ptr<T: 'static>(
&self,
key: &'static UserDataKey<T>
) -> Option<NonNull<T>>

Return the user data previously attached to self with the given key, if any, without incrementing the reference count.

The pointer is valid when it is returned from this method, until the cairo object that self represents is destroyed or remove_user_data or set_user_data is called with the same key.

pub fn remove_user_data<T: 'static>(
&self,
key: &'static UserDataKey<T>
) -> Result<(), Error>

Unattached from self the user data associated with key, if any. If there is no other Rc strong reference, the data is destroyed.