Struct cairo::ScaledFont

pub struct ScaledFont(_);

Implementations

impl ScaledFont

pub fn new(
font_face: &FontFace,
font_matrix: &Matrix,
ctm: &Matrix,
options: &FontOptions
) -> ScaledFont

pub fn to_raw_none(&self) -> *mut cairo_scaled_font_t

pub unsafe fn from_raw_full(ptr: *mut cairo_scaled_font_t) -> ScaledFont

pub unsafe fn from_raw_none(ptr: *mut cairo_scaled_font_t) -> ScaledFont

pub fn get_type(&self) -> FontType

pub fn get_reference_count(&self) -> usize

pub fn extents(&self) -> FontExtents

pub fn text_extents(&self, text: &str) -> TextExtents

pub fn glyph_extents(&self, glyphs: &[Glyph]) -> TextExtents

pub fn text_to_glyphs(
&self,
x: f64,
y: f64,
text: &str
) -> (Vec<Glyph>, Vec<TextCluster>)

pub fn get_font_face(&self) -> FontFace

pub fn get_font_options(&self) -> FontOptions

pub fn get_font_matrix(&self) -> Matrix

pub fn get_ctm(&self) -> Matrix

pub fn get_scale_matrix(&self) -> Matrix

pub fn set_user_data<T: 'static>(
&self,
key: &'static UserDataKey<T>,
value: Rc<T>
)

Attach user data to self for the given key.

pub fn get_user_data<T: 'static>(
&self,
key: &'static UserDataKey<T>
) -> Option<Rc<T>>

Return the user data previously attached to self with the given key, if any.

pub fn get_user_data_ptr<T: 'static>(
&self,
key: &'static UserDataKey<T>
) -> Option<NonNull<T>>

Return the user data previously attached to self with the given key, if any, without incrementing the reference count.

The pointer is valid when it is returned from this method, until the cairo object that self represents is destroyed or remove_user_data or set_user_data is called with the same key.

pub fn remove_user_data<T: 'static>(&self, key: &'static UserDataKey<T>)

Unattach from self the user data associated with key, if any. If there is no other Rc strong reference, the data is destroyed.